move RegistrationView into binder package
