+++ apple/OmnivoreKit/Sources/Models/AuthFlow.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum AuthFlow {
+  case signIn
+  case signUp
+}
