+++ apple/OmnivoreKit/Sources/Views/RegistrationViews/ToggleAuthFlowButton.swift
@@ -1,3 +1,4 @@
+import Models
 import SwiftUI
 
 struct ToggleAuthFlowButton: View {
